+++ wgpu-core/src/command/render.rs
@@ -1189,7 +1189,15 @@ impl RenderPassInfo {
                     .flags
                     .contains(wgt::DownlevelFlags::READ_ONLY_DEPTH_STENCIL)
             {
-                wgt::TextureUses::DEPTH_STENCIL_READ | wgt::TextureUses::RESOURCE
+                // If the texture supports TEXTURE_BINDING, it can be used as a shader
+                // resource and a read-only depth attachment simultaneously. But if it
+                // doesn't support TEXTURE_BINDING, don't attempt to transition it to a
+                // shader resource state, because DX12 will raise an error.
+                if view.desc.usage.contains(TextureUsages::TEXTURE_BINDING) {
+                    wgt::TextureUses::DEPTH_STENCIL_READ | wgt::TextureUses::RESOURCE
+                } else {
+                    wgt::TextureUses::DEPTH_STENCIL_READ
+                }
             } else {
                 wgt::TextureUses::DEPTH_STENCIL_WRITE
             };
@@ -2791,7 +2799,7 @@ fn multi_draw_indirect(
         return Err(RenderPassErrorInner::UnalignedIndirectBufferOffset(offset));
     }
 
-    let stride = get_stride_of_indirect_args(family);
+    let stride = get_src_stride_of_indirect_args(family);
 
     let end_offset = offset + stride * count as u64;
     if end_offset > indirect_buffer.size {
@@ -2909,9 +2917,12 @@ fn multi_draw_indirect(
             let draw_data = draw_ctx.add(offset + stride * i as u64)?;
 
             if draw_data.buffer_index == current_draw_data.buffer_index {
+                #[cfg(debug_assertions)]
+                let dst_stride =
+                    get_dst_stride_of_indirect_args(state.pass.base.device.backend(), family);
                 debug_assert_eq!(
                     draw_data.offset,
-                    current_draw_data.offset + stride * current_draw_data.count as u64
+                    current_draw_data.offset + dst_stride * current_draw_data.count as u64
                 );
                 current_draw_data.count += 1;
             } else {
@@ -2963,7 +2974,7 @@ fn multi_draw_indirect_count(
         validate_mesh_draw_multiview(state)?;
     }
 
-    let stride = get_stride_of_indirect_args(family);
+    let stride = get_src_stride_of_indirect_args(family);
 
     state
         .pass
@@ -3832,10 +3843,23 @@ impl Global {
     }
 }
 
-pub(crate) const fn get_stride_of_indirect_args(family: DrawCommandFamily) -> u64 {
+pub(crate) const fn get_src_stride_of_indirect_args(family: DrawCommandFamily) -> u64 {
     match family {
         DrawCommandFamily::Draw => size_of::<wgt::DrawIndirectArgs>() as u64,
         DrawCommandFamily::DrawIndexed => size_of::<wgt::DrawIndexedIndirectArgs>() as u64,
         DrawCommandFamily::DrawMeshTasks => size_of::<wgt::DispatchIndirectArgs>() as u64,
     }
+}
+
+pub(crate) const fn get_dst_stride_of_indirect_args(
+    backend: wgt::Backend,
+    family: DrawCommandFamily,
+) -> u64 {
+    // space for D3D12 special constants
+    let extra = if matches!(backend, wgt::Backend::Dx12) {
+        3 * size_of::<u32>() as u64
+    } else {
+        0
+    };
+    extra + get_src_stride_of_indirect_args(family)
 }